// src/modules/FileReadsViewer/types/table-body.type.ts
export type FileReadsTableBody = {
  type: string,
  fileId: string
  filePath: string
  fileVersion: string
  processVersionId: string,
  process: string
  processVersion: string
  user: string
  uid: string
  originFile: string
  filesystem: string
  depth: string
  firstAt: string
  lastAt: string
  count: string
  initialSizeBytes: string
  trackingStartAt: string
  deletedAt: string
}