// src/modules/FileReadsViewer/types/message-event.dto.ts
export interface MessageEventDto {
  status?: string
  filesystemId?: string
  trackingStartedAt?: string
  birthTime?: string
  fileType?: string
  versionNumber?: number
  osUserId?: string
  executablePath?: string
  operationType?: string
  firstAt?: string
  limit?: number
  page?: number
  process?: string
}